fix: bugs after merge.
