+++ packages/utilities/src/lib/helpers/index.ts
@@ -29,7 +29,7 @@ export const createMarkdocFunction = (cb: Callback) => {
     }
 
 };
-};
+
 
 
 
@@ -329,4 +329,3 @@ export async function returnErrorAndResultFromPromise<T extends Promise<any>>(pr
 
 };
 
-export { throwIf, throwUnless, returnErrorAndResultFromPromise }
